+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 ## 6.1.3 (TBD)
 
 * UrlTileGenerator HTTP Method and Header field values support
+* TileCreator improved tile sizes when tile width and/or height are not specified
 * gradle plugin updated to 7.0.4
 
 ## [6.1.2](https://github.com/ngageoint/geopackage-android/releases/tag/6.1.2) (11-22-2021)